moved audio dir into src and changed path in meditate breathing and reflect component.
